reorganize some things
 * move flopy4/mf6/dimensions.py to flopy4/dimensions.py (not mf6-specific)
 * relocate test_mf6_*.py files into test/mf6/ with consistent naming
 * add demo_*.py to pytest discovery and rename some tests as demos
