+++ flopy4/mf6/component.py
@@ -9,9 +9,9 @@ from xattree import asdict as xattree_asdict
 from xattree import xattree
 
 from flopy4.mf6.constants import MF6
-from flopy4.mf6.dimensions import DimensionResolverMixin
-from flopy4.mf6.spec import fields_dict
 from flopy4.mf6.spec import xattree_field as field
+from flopy4.dimensions import DimensionResolverMixin
+from flopy4.mf6.spec import field, fields_dict
 from flopy4.mf6.utils.grid import update_maxbound
 from flopy4.mf6.write_context import WriteContext
 from flopy4.uio import IO, Loader, Writer
