+++ pytest.ini
@@ -3,5 +3,6 @@ addopts = -ra --color=yes --benchmark-disable
 testpaths = test
 python_files =
     test_*.py
+    demo_*.py
 markers =
     slow: tests not completing in a few seconds
