+++ test/demo_xarray_networkx_disu.py
