+++ test/mf6/test_mf6_netcdf.py
