+++ test/mf6/test_mf6_netcdf_cf_conventions.py
